requirements.txt
# ── Acne Flare Predictor — Python Requirements ──────────────────────────────
# Python 3.9+  |  GPU strongly recommended for training (CPU works for inference)

# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data & preprocessing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Weather API
openmeteo-requests>=1.1.0
requests-cache>=1.1.0
retry-requests>=2.0.0
requests>=2.31.0

# Explainability
shap>=0.43.0

# Streamlit frontend
streamlit>=1.35.0
plotly>=5.18.0

# Notebook (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.7.0
